agentic_code_review_assistant/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
langchain
openai
qdrant-client
numpy
pytest
httpx
pydantic
gitpython
tqdm
